+++ src/pages/questionnaire/politicalCompassTest.js
@@ -1,0 +1,116 @@
+export const politicalCompassTest = [
+  {
+    id: 1,
+    questionText:
+      'If economic globalisation is inevitable, it should primarily serve humanity rather than the interests of trans-national corporations.',
+    answers: [
+      {
+        text: 'Strongly disagree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Disagree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Agree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Strongly agree',
+        points: { x: 1, y: 1 },
+      },
+    ],
+  },
+  {
+    id: 2,
+    questionText:
+      'I’d always support my country, whether it was right or wrong.',
+    answers: [
+      {
+        text: 'Strongly disagree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Disagree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Agree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Strongly agree',
+        points: { x: 1, y: 1 },
+      },
+    ],
+  },
+  {
+    id: 3,
+    questionText:
+      'No one chooses his or her country of birth, so it’s foolish to be proud of it.',
+    answers: [
+      {
+        text: 'Strongly disagree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Disagree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Agree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Strongly agree',
+        points: { x: 1, y: 1 },
+      },
+    ],
+  },
+  {
+    id: 4,
+    questionText:
+      'Our race has many superior qualities, compared with other races.',
+    answers: [
+      {
+        text: 'Strongly disagree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Disagree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Agree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Strongly agree',
+        points: { x: 1, y: 1 },
+      },
+    ],
+  },
+  {
+    id: 5,
+    questionText: 'The enemy of my enemy is my friend.',
+    answers: [
+      {
+        text: 'Strongly disagree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Disagree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Agree',
+        points: { x: 1, y: 1 },
+      },
+      {
+        text: 'Strongly agree',
+        points: { x: 1, y: 1 },
+      },
+    ],
+  },
+]
